#7 It works, but it isn't actually useful
The problem is that bouncer doesn't seem really to support the idea that an object may satisfy one of a set of validations, and that's OK; and it doesn't seem to support collecting nested errors through custom validators.

So although this will now tell you *that* an adl structure is invalid, it won't tell you *where* the document is invalid.